+++ src/main/java/dat3/aym/service/ProductService.java
@@ -1,8 +1,8 @@
-package dat3.product.service;
+package dat3.aym.service;
 
-import dat3.product.dto.ProductDto;
-import dat3.product.entity.Product;
-import dat3.product.repository.ProductRepository;
+import dat3.aym.dto.ProductDto;
+import dat3.aym.entity.Product;
+import dat3.aym.repository.ProductRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -10,7 +10,6 @@ import org.springframework.stereotype.Service;
 import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
-import java.util.Optional;
 
 @Service
 public class ProductService {
